Убраны сеттеры из dto, валидация в сервисе
