+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -1,9 +1,7 @@
 package ru.practicum.shareit.booking.dto;
 
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
 public class BookingDto {
 }
