+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -2,11 +2,9 @@ package ru.practicum.shareit.item.dto;
 
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 import ru.practicum.shareit.request.ItemRequest;
 
 @Getter
-@Setter
 @Builder
 public class ItemDto {
     private Long id;
