+++ src/main/java/ru/practicum/shareit/item/repository/ItemRepositoryImpl.java
@@ -1,8 +1,6 @@
 package ru.practicum.shareit.item.repository;
 
 import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.exception.ConditionsNotMetException;
-import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.item.model.Item;
 
 import java.util.HashMap;
@@ -13,14 +11,12 @@ import java.util.stream.Collectors;
 
 @Repository
 public class ItemRepositoryImpl implements ItemRepository {
-    private static final String ITEM_NOT_FOUND = "Вещь с id = %d не найдена";
     private final Map<Long, Item> items = new HashMap<>();
 
     private long newId = 0;
 
     @Override
     public Item saveItem(Item item, Long ownerId) {
-        validateCreate(item);
         item.setId(++newId);
         item.setOwnerId(ownerId);
         items.put(item.getId(), item);
@@ -28,12 +24,8 @@ public class ItemRepositoryImpl implements ItemRepository {
     }
 
     @Override
-    public Item updateItem(Long itemId, Item item, Long userId) {
-        Item itemFromRepository = getItemById(itemId)
-                .orElseThrow(() -> new NotFoundException(String.format(ITEM_NOT_FOUND, item.getId())));
-
-        validateUserIsOwner(itemFromRepository, userId);
-
+    public Item updateItem(Long itemId, Item item) {
+        Item itemFromRepository = getItemById(itemId).get(); // проверка в сервисе
         if (item.getName() != null) {
             itemFromRepository.setName(item.getName());
         }
@@ -59,10 +51,7 @@ public class ItemRepositoryImpl implements ItemRepository {
     }
 
     @Override
-    public void deleteItemById(Long id, Long userId) {
-        Item item = getItemById(id)
-                .orElseThrow(() -> new NotFoundException(String.format(ITEM_NOT_FOUND, id)));
-        validateUserIsOwner(item, userId);
+    public void deleteItemById(Long id) {
         items.remove(id);
     }
 
@@ -76,23 +65,5 @@ public class ItemRepositoryImpl implements ItemRepository {
                         item.getName().toLowerCase().contains(searchText) ||
                         item.getDescription().toLowerCase().contains(searchText))
                 .collect(Collectors.toList());
-    }
-
-    private void validateCreate(Item item) {
-        if (item.getName().isBlank()) {
-            throw new ConditionsNotMetException("Название должно быть указано");
-        }
-        if (item.getDescription().isBlank()) {
-            throw new ConditionsNotMetException("Описание должно быть указано");
-        }
-        if (item.isAvailable() == null) {
-            throw new ConditionsNotMetException("Доступность должна быть указана");
-        }
-    }
-
-    private void validateUserIsOwner(Item item, Long userId) {
-        if (!item.getOwnerId().equals(userId)) {
-            throw new ConditionsNotMetException("Редактировать или удалять вещь может только её владелец");
-        }
     }
 }
