+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -2,6 +2,7 @@ package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.ConditionsNotMetException;
 import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.mapper.ItemMapper;
@@ -20,13 +21,17 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public ItemDto saveItem(ItemDto item, Long ownerId) {
+        validateCreate(item);
         Item savedItem = itemRepository.saveItem(ItemMapper.toItem(item), ownerId);
         return ItemMapper.toItemDto(savedItem);
     }
 
     @Override
     public ItemDto updateItem(Long itemId, Item item, Long userId) {
-        Item updatedItem = itemRepository.updateItem(itemId, item, userId);
+        Item itemFromRepository = itemRepository.getItemById(itemId)
+                .orElseThrow(() -> new NotFoundException(String.format(ITEM_NOT_FOUND, item.getId())));
+        validateUserIsOwner(itemFromRepository, userId);
+        Item updatedItem = itemRepository.updateItem(itemId, item);
         return ItemMapper.toItemDto(updatedItem);
     }
 
@@ -46,7 +51,10 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public void deleteItemById(Long itemId, Long userId) {
-        itemRepository.deleteItemById(itemId, userId);
+        Item item = itemRepository.getItemById(itemId)
+                .orElseThrow(() -> new NotFoundException(String.format(ITEM_NOT_FOUND, itemId)));
+        validateUserIsOwner(item, userId);
+        itemRepository.deleteItemById(itemId);
     }
 
     @Override
@@ -57,5 +65,23 @@ public class ItemServiceImpl implements ItemService {
         return itemRepository.searchByText(text).stream()
                 .map(ItemMapper::toItemDto)
                 .collect(Collectors.toList());
+    }
+
+    private void validateCreate(ItemDto item) {
+        if (item.getName().isBlank()) {
+            throw new ConditionsNotMetException("Название должно быть указано");
+        }
+        if (item.getDescription().isBlank()) {
+            throw new ConditionsNotMetException("Описание должно быть указано");
+        }
+        if (item.isAvailable() == null) {
+            throw new ConditionsNotMetException("Доступность должна быть указана");
+        }
+    }
+
+    private void validateUserIsOwner(Item item, Long userId) {
+        if (!item.getOwnerId().equals(userId)) {
+            throw new ConditionsNotMetException("Редактировать или удалять вещь может только её владелец");
+        }
     }
 }
