+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -2,10 +2,8 @@ package ru.practicum.shareit.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
 @AllArgsConstructor
 public class UserDto {
     private Long id;
